Refactor API key management: move generate and revoke functions to helpers and update URLs
